cleaning up tests
